final UML (almost) + modifs cohérence code/UML

+++ src/main/java/fr/univamu/iut/marche/traitement/acteurs/Offre.java
@@ -4,12 +4,12 @@ import fr.univamu.iut.marche.traitement.produits.ProduitFermier;
 
 public class Offre {
     private Participant.Produits produitOffre;
-    private Integer prixOffre;
+    private double prixOffre;
     private Participant acheteur;
     private Marche marche;
     private Integer quantite;
 
-    public Offre(Participant.Produits produitOffre, Participant acheteur, Integer prixOffre,Integer quantite, Marche marche) {
+    public Offre(Participant.Produits produitOffre, Participant acheteur, double prixOffre,Integer quantite, Marche marche) {
         this.produitOffre = produitOffre;
         this.prixOffre = prixOffre;
         this.acheteur = acheteur;
@@ -41,11 +41,12 @@ public class Offre {
     public void setMarche(Marche marche) {
         this.marche = marche;
     }
-    public Integer getPrixOffre() {
+
+    public double getPrixOffre() {
         return prixOffre;
     }
 
-    public void setPrixOffre(Integer prixOffre) {
+    public void setPrixOffre(double prixOffre) {
         this.prixOffre = prixOffre;
     }
